rename html_image_codec to html_image_element_codec
