+++ lib/web_ui/lib/src/engine.dart
@@ -106,7 +106,7 @@ export 'engine/html/shaders/vertex_shaders.dart';
 export 'engine/html/surface.dart';
 export 'engine/html/surface_stats.dart';
 export 'engine/html/transform.dart';
-export 'engine/html_image_codec.dart';
+export 'engine/html_image_element_codec.dart';
 export 'engine/image_decoder.dart';
 export 'engine/initialization.dart';
 export 'engine/js_interop/js_app.dart';
